llamaindex_agents/requirements.txt
# 依赖包
python-dotenv
openai
fastapi
uvicorn
llama-index
